Create conventional project file structure
