+++ cmd/web/main.go
@@ -1,0 +1,18 @@
+package main
+
+import (
+	"log"
+	"net/http"
+)
+
+// Router
+func main() {
+	mux := http.NewServeMux()
+	mux.HandleFunc("/", home)
+	mux.HandleFunc("/snippet/view", snippetView)
+	mux.HandleFunc("/snippet/create", snippetCreate)
+
+	log.Print("Starting server on :4000")
+	err := http.ListenAndServe(":4000", mux)
+	log.Fatal(err)
+}
